Add Scores voting method.
