+++ .idea/multipoll.iml
@@ -17,13 +17,10 @@
       <excludeFolder url="file://$MODULE_DIR$/.venv" />
       <excludeFolder url="file://$MODULE_DIR$/multipoll/migrations" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.7.5 virtualenv at ~/TProj/Python/multipoll/.venv" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.7 (multipoll)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
   <component name="TemplatesService">
     <option name="TEMPLATE_CONFIGURATION" value="Django" />
-  </component>
-  <component name="TestRunnerService">
-    <option name="PROJECT_TEST_RUNNER" value="Unittests" />
   </component>
 </module>
