+++ multipoll/electoralsystems/__init__.py
@@ -1,9 +1,6 @@
 from multipoll.electoralsystems.approval import approval
 from multipoll.electoralsystems.borda import borda
 from multipoll.electoralsystems.rankedpairs import ranked_pairs
-from multipoll.electoralsystems.ranking import Ranking
-from multipoll.electoralsystems.registry import electoral_system, get_electoral_system
+from multipoll.electoralsystems.utils import get_electoral_system
 
-
-__all__ = ["electoral_system", "Ranking", "get_electoral_system",
-           "approval", "borda", "ranked_pairs"]
+__all__ = ["get_electoral_system", "approval", "borda", "ranked_pairs"]
