+++ multipoll/electoralsystems/utils/registry.py
