+++ multipoll/models/multipoll.py
@@ -1,7 +1,7 @@
 from __future__ import annotations  # noqa: T499
 # noqa: T499, E800
 
-from typing import Dict, Type
+from typing import ClassVar, Dict, Tuple, Type
 
 from django import forms
 from django.db import models
@@ -10,11 +10,11 @@ from multipoll.models.pollbase import FullVoteBase, PartialVoteBase, PollBase
 
 
 class MultiPoll(PollBase):
-    WeightFieldType: Type['models.SmallIntegerField[int, int]'] = models.SmallIntegerField
-    weight_field_args = {'null': True}
+    WeightFieldType: ClassVar[Type['models.SmallIntegerField[int, int]']] = models.SmallIntegerField
+    weight_field_args: ClassVar[Dict] = {'null': True}
 
-    supported_systems = ("approval", "borda", "rankedpairs")
-    default_system = "rankedpairs"
+    supported_systems: ClassVar[Tuple[str, ...]] = ("approval", "borda", "rankedpairs", "score")
+    default_system: ClassVar[str] = "rankedpairs"
 
     def create_attachment_for_option(self, ind: int) -> Dict[str, str]:
         attach = {"name": "int_option", "text": self.options[ind], "type": "button",
